silenced warning.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@4537 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/localnav/testsim/simulator.cpp
@@ -189,7 +189,7 @@ Simulator::time() const
 
     orca::Time t;
     t.seconds = (int)sec;
-    t.useconds = (sec-(int)sec) * 1e6;
+    t.useconds = (int)((sec-(int)sec) * 1e6);
 
     return t;
 }
